extract logic of submitting outfit payload to backend
