+++ src/container/mainWrapper/mainWrapper.tsx
@@ -1,20 +1,9 @@
 import * as React from 'react';
 import {
-  IClothing,
-  IOutfit
+  IMainContext
 } from '../../App';
 import Main from '../main/main';
 import { MainContext } from '../mainProvider/mainProvider';
-
-
-export interface IMainContext {
-  clothings : IClothing[],
-  outfits   : IOutfit[],
-  headwear  : IClothing[];
-  top       : IClothing[];
-  bottom    : IClothing[];
-  shoes     : IClothing[];
-}
 
 class MainWrapper extends React.PureComponent < {}, {} > {
   public render(){
